Narrative Engine updates
Updated narrative engine with a separated out ChatLogBuilder
The ChatLogBuilder now manages parsing a chatlog and dumps it into a giant list of data. This isn't remotely close to complete yet but it's a good start - we still need to build out a spreadsheet for the character information so that we can start parsing and building those data setups as well as updating how we parse this information in the chatlog to link to that data.

Additionally added several plugins, such as 'NaughtyAttributes'.

Began a bit of code cleanup.