Vending Machine Updates
Door works now.
Item delivery implemented.
Once items are picked up from the delivery tray the door closes.
Consider adding a trigger collider that works to ensure the door remains open (and the vending machine remains locked for use) when items are in the bay.

Possible space for narrative/gameplay?